refractored tempCleanup in challenge model

+++ app/models/challenge.rb
@@ -64,19 +64,16 @@ class Challenge < ActiveRecord::Base
 
   def tempCleanUp(file_path)
     #Path parameters
-    zip_path = file_path + 'tmp'+ self.id.to_s + '.zip'
-    yaml_path = file_path + 'tmp' + self.id.to_s + '.yml'
+    # zip_path = file_path + 'tmp'+ self.id.to_s + '.zip'
+    # yaml_path = file_path + 'tmp' + self.id.to_s + '.yml'
 
     #Delete html files
     self.webpages.each do |webpage|
-      if webpage.title != ""
-        html_name = webpage.title + '.html'
-        File.delete(file_path + html_name)
-      end
+      html_name = webpages.title + '.html' and File.delete(file_path + html_name) if not webpage.title.empty?
     end
 
-    File.delete(zip_path) #Delete zip file
-    File.delete(yaml_path) #Delete yaml file
+    File.delete(file_path + 'tmp'+ self.id.to_s + '.zip') #Delete zip file
+    File.delete(file_path + 'tmp' + self.id.to_s + '.yml') #Delete yaml file
   end
 
 end
